+++ src/types/requests/update-subject-request.model.ts
@@ -1,3 +1,4 @@
 export interface UpdateSubjectRequest {
   name?: string;
+  imageUrl?: string | null;
 }