+++ src/types/responses/subject-reponse.model.ts
@@ -2,6 +2,7 @@ export interface SubjectResponse {
   id: number;
   name: string;
   userId: string;
+  imageUrl?: string | null;
   professor?: {
     name: string;
     lastName: string;